requirements.txt
Django
django-cors-middleware
django-cors-headers
django-crispy-forms
django-docs
django-extensions
django-filter
django-rest-swagger
djangorestframework
djangorestframework-jwt
python-decouple
psycopg2
django-rest-swagger

